Move Avatar tooltip container to MemberInfo
